Convert HWRP to a custom resource
